requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0
lxml>=4.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pyyaml>=6.0.0
typing-extensions>=4.0.0
aiohttp>=3.8.0
python-dateutil>=2.8.0
